backend/requirements.txt
# Django
Django==1.11

# Django Channels for Websockets
channels==1.1.3

# REST Framework
djangorestframework==3.6.2

# Documenting the REST API
coreapi==2.3.0

# Postgres
psycopg2==2.7.1
